% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowling.R
\docType{data}
\name{bowling}
\alias{bowling}
\title{Bowling ratings.}
\format{A data frame with 1,784,977 rows and 5 variables:
\describe{
\item{Date}{Date.}
\item{Format}{The type of cricket.}
\item{Rank}{Ranking.}
\item{Rating}{ICC player rating.}
\item{Name}{The name of the player.}
}}
\source{
\url{http://www.relianceiccrankings.com/}
}
\usage{
bowling
}
\description{
A dataset containing bowling ratings by date since 1st January 2000.
}
\keyword{datasets}
