% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_batting}
\alias{scrape_batting}
\title{Scrape the top 100 batting ratings by format for a given date.}
\usage{
scrape_batting(format, date)
}
\arguments{
\item{format}{One of 'Test', 'ODI' or 'T20'.}

\item{date}{The date to look up the rankings for.}
}
\value{
A tibble with a row for each player.

The fields are:

\describe{
\item{Date}{The date to which these rankings and ratings apply.}
\item{Format}{The type of cricket.}
\item{Rank}{Ranking.}
\item{Rating}{ICC player rating.}
\item{Name}{The name of the player.}
}
}
\description{
Scrape the top 100 batting ratings by format for a given date.
}
